DHFPROD-4519: Add AppInfo test
