+++ one-ui/src/test/java/com/marklogic/hub/oneui/controllers/AppInfoTest.java
@@ -1,0 +1,51 @@
+/*
+ * Copyright 2012-2020 MarkLogic Corporation
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ */
+package com.marklogic.hub.oneui.controllers;
+
+import com.marklogic.hub.oneui.Application;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.core.env.Environment;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.springframework.test.context.web.WebAppConfiguration;
+
+import java.util.Objects;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+@ExtendWith(SpringExtension.class)
+@SpringBootTest(classes = {Application.class})
+@WebAppConfiguration
+public class AppInfoTest {
+
+    @Autowired
+    AppInfoController appInfoController;
+
+    @Autowired
+    Environment environment;
+
+    @Test
+    public void getInfo() {
+        String expectedTimeout = environment.getProperty("server.servlet.session.timeout");
+        String actualTimeout = Objects.requireNonNull(appInfoController.getInfo().getBody())
+            .get("session.timeout");
+        assertEquals(expectedTimeout, actualTimeout);
+    }
+
+}
